+++ recipes/api-metadata.rb
@@ -47,7 +47,7 @@ service 'nova-api-metadata' do
   supports status: true, restart: true
   subscribes :restart, resources('template[/etc/nova/nova.conf]')
 
-  action :enable
+  action [:enable, :start]
 end
 
 template '/etc/nova/api-paste.ini' do